# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin-mall API Documentation",
        "title": "gin-mall",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "1478488313@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/avatar": {
            "post": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "上传头像，文件类型支持jpg(jpeg)、png",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "上传用户头像",
                "parameters": [
                    {
                        "type": "file",
                        "description": "头像",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UploadAvatarResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/email/bind": {
            "post": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "发送邮件到用户指定邮箱，用户确认后进行绑定",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "绑定邮箱",
                "parameters": [
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.BindEmailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/email/valid": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "通过指定链接确认绑定操作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "邮箱绑定确认",
                "parameters": [
                    {
                        "type": "string",
                        "description": "email token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ValidEmailResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/follow": {
            "post": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "基于id关注的其他用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "关注用户",
                "parameters": [
                    {
                        "description": "关注用户的id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserFollowResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "基于id取关用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "取关用户",
                "parameters": [
                    {
                        "description": "取关用户的id",
                        "name": "id",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserUnfollowResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/follower": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "获取用户的粉丝列表，列表中的用户信息包括头像和昵称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取粉丝列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserFollowerListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/following": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "获取用户的关注列表，列表中的用户信息包括头像和昵称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取关注列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserFollowingListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "可查看信息包括：用户名、昵称、用户状态、邮箱、头像",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "查看用户信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.ShowUserInfoResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "可修改用户昵称",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "昵称",
                        "name": "nickname",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserInfoUpdateResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/password": {
            "put": {
                "security": [
                    {
                        "AccessToken": []
                    },
                    {
                        "RefreshToken": []
                    }
                ],
                "description": "输入`原密码-新密码-确认密码`以更改密码",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "更改用户密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "原密码",
                        "name": "origin_password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新密码",
                        "name": "new_password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "确认密码",
                        "name": "confirm_password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.UserPasswordChangeResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/e.Code"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "e.Code": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15,
                16,
                17,
                18,
                19,
                20,
                21,
                22,
                23,
                24,
                25,
                26,
                27,
                28,
                29,
                30
            ],
            "x-enum-comments": {
                "ErrorAccountInvalid": "用户名或密码错误",
                "ErrorContextValue": "上下文值传递错误",
                "ErrorCreateUser": "创建用户错误",
                "ErrorEmailLinkExpire": "邮件确认链接已过期",
                "ErrorEncryptMoney": "金额加密错误",
                "ErrorEncryptPassword": "密码加密错误",
                "ErrorFileError": "文件错误",
                "ErrorFileType": "文件类型错误",
                "ErrorFollowUser": "关注用户失败",
                "ErrorGenerateToken": "token生成错误",
                "ErrorGetCategoryList": "获取商品分类失败",
                "ErrorGetFollowerList": "获取粉丝列表失败",
                "ErrorGetFollowingList": "获取关注列表失败",
                "ErrorGetProductByID": "根据ID获取商品失败",
                "ErrorGetProductList": "获取商品列表失败",
                "ErrorGetUserByID": "根据id获取用户失败",
                "ErrorIncorrectPassword": "密码错误",
                "ErrorInvalidIDParam": "非法的id参数",
                "ErrorOSSUploadError": "OSS文件上传错误",
                "ErrorParseToken": "token解析错误",
                "ErrorSendEmail": "发送邮件错误",
                "ErrorSendEmailTooFrequent": "邮件发送操作频繁",
                "ErrorUnfollowUser": "取消关注失败",
                "ErrorUpdateEmail": "更新邮箱错误",
                "ErrorUpdateUser": "更新用户失败",
                "ErrorUploadAvatar": "头像上传错误",
                "ErrorUploadFile": "文件上传错误",
                "ErrorUserExists": "用户已存在",
                "InvalidParams": "参数错误",
                "Success": "响应成功",
                "UnknownError": "未知错误"
            },
            "x-enum-varnames": [
                "Success",
                "InvalidParams",
                "UnknownError",
                "ErrorUserExists",
                "ErrorEncryptPassword",
                "ErrorEncryptMoney",
                "ErrorCreateUser",
                "ErrorAccountInvalid",
                "ErrorGetUserByID",
                "ErrorUpdateUser",
                "ErrorIncorrectPassword",
                "ErrorUploadAvatar",
                "ErrorFollowUser",
                "ErrorUnfollowUser",
                "ErrorGetFollowingList",
                "ErrorGetFollowerList",
                "ErrorGenerateToken",
                "ErrorParseToken",
                "ErrorContextValue",
                "ErrorUploadFile",
                "ErrorFileError",
                "ErrorOSSUploadError",
                "ErrorFileType",
                "ErrorSendEmail",
                "ErrorUpdateEmail",
                "ErrorEmailLinkExpire",
                "ErrorSendEmailTooFrequent",
                "ErrorGetCategoryList",
                "ErrorGetProductList",
                "ErrorGetProductByID",
                "ErrorInvalidIDParam"
            ]
        },
...
        "response.BindEmailResp": {
            "type": "object"
        },
...
        "response.Follow": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "response.ShowUserInfoResp": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "response.UploadAvatarResp": {
            "type": "object"
        },
        "response.UserFollowResp": {
            "type": "object"
        },
        "response.UserFollowerListResp": {
            "type": "object",
            "properties": {
                "follower": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Follow"
                    }
                }
            }
        },
        "response.UserFollowingListResp": {
            "type": "object",
            "properties": {
                "following": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Follow"
                    }
                }
            }
        },
        "response.UserInfoUpdateResp": {
            "type": "object"
        },
        "response.UserPasswordChangeResp": {
            "type": "object"
        },
        "response.UserUnfollowResp": {
            "type": "object"
        },
        "response.ValidEmailResp": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "AccessToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "RefreshToken": {
            "type": "apiKey",
            "name": "X-Refresh-Token",
            "in": "header"
        }
    },
...
}